Move NormalDotFlux from Evolution to DG
