+++ tests/Unit/NumericalAlgorithms/DiscontinuousGalerkin/CMakeLists.txt
@@ -6,6 +6,7 @@ set(LIBRARY "Test_NumericalDiscontinuousGalerkin")
 set(LIBRARY_SOURCES
   Test_LiftFlux.cpp
   Test_MortarHelpers.cpp
+  Test_NormalDotFlux.cpp
   Test_SimpleBoundaryData.cpp
   Test_Tags.cpp
   )
